Refactor: Separate MCP tools by category (#8)
