+++ README_COLLAB.md
@@ -1,0 +1,4 @@
+# ScaffoldKit: MCP Spring DevTools Collaborator's Guide 🏗️
+
+
+
